SDI-477 Mark messages in error to retry after new queue property `errorVisibilityTimeout`
