+++ hmpps-sqs-spring-boot-autoconfigure/src/main/kotlin/uk/gov/justice/hmpps/sqs/HmppsSqsProperties.kt
@@ -21,6 +21,7 @@ data class HmppsSqsProperties(
     val dlqSecretAccessKey: String = "",
     val dlqMaxReceiveCount: Int = 5,
     val visibilityTimeout: Int = 30,
+    val errorVisibilityTimeout: Int = 0,
   )
 
   data class TopicConfig(
