+++ test-app/src/test/resources/application-test.yml
@@ -14,7 +14,7 @@ hmpps.sqs:
       subscribeTopicId: inboundtopic
       subscribeFilter: '{"eventType":[ "OFFENDER_MOVEMENT-RECEPTION", "OFFENDER_MOVEMENT-DISCHARGE", "test.type"] }'
       dlqMaxReceiveCount: 1
-      visibilityTimeout: 1
+      errorVisibilityTimeout: 0
     outboundqueue:
       dlqName: ${random.uuid}
       queueName: ${random.uuid}
